NMAP Menu
Quick and dirty menu that prompts a user for input to run host
detection, port, and service enumeration on a network.  can be run on a
single system or a range of systems using 192.168.1.0-50 format or
traditional cidr.  Outputs the results of scans in separate files in the
local directory.